requirements.txt
fastapi
google-auth
httpx
pydantic-settings
requests
uvicorn
